refactor: 코드리뷰 반영 - request dto 패키지명 오타 - Review 내 method naming 규칙 일관성x - ReviewImagePolicy static 키워드 누락 - tag 가져오는 과정에서의 잠재적 오류 해결
