+++ src/main/java/com/ururulab/ururu/global/controller/Controller.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.global.Controller;
+package com.ururulab.ururu.global.controller;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
